Switched rating from Avg to Sum

+++ src/main/java/hs/aalen/urlaub/vacationWish/VacationWishAndRating.java
@@ -4,18 +4,18 @@ package hs.aalen.urlaub.vacationWish;
 //This class is used to display the average rating of a vacation wish
 //It exists to overcome an error that occurs when trying to receive Ratings from the frontend
 public class VacationWishAndRating extends VacationWish {
-    private Double averageRating;
+    private int averageRating;
 
     public VacationWishAndRating() {
         super();
     }
 
-    public VacationWishAndRating(VacationWish vacationWish, Double averageRating) {
+    public VacationWishAndRating(VacationWish vacationWish, int averageRating) {
         super(vacationWish.getId(), vacationWish.getLocation(), vacationWish.getDescription());
         this.averageRating = averageRating;
     }
 
-    public Double getAverageRating() {
+    public int getAverageRating() {
         return averageRating;
     }
 }
